requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.5
django-allauth==0.42.0
django-cors-headers==3.4.0
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
djongo==1.3.3
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
lock==2018.3.25.2110
MarkupSafe==1.1.1
oauthlib==3.1.0
pipfile==0.0.2
psycopg2==2.8.5
pymongo==3.11.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.2.4
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
